Create AS09, fallout hacking terminal
In the game Fallout 4 you are expected to crack codes. It is similar to hangman. The game includes a word list and tells you a likeness compared to the actual word. For it to be like the actual word not only must the letter be in the word, but also in the same position. To do this in Python you will need to type in the word and check it.