Create AS26, word length calculator
A rough guide as to how difficult a document is to read is to work out word lengths. Write a program that will calculate the average word length of a text stored in a file (i.e the sum of all the lengths of the word tokens in the text, divided by the number of word tokens).